Improve digimon queries, fix evolution logic, and enhance null safety
Added pagination to GetAllCharactersDigimonQuery and its handler. Refactored GameServer disconnect logic for better null safety and error handling. Updated evolution and encyclopedia checking logic for performance and reliability. Fixed evolution XG/DS values in PartnerEvolutionPacketProcessor. Improved mapping and DTOs for null safety. Removed unused ConsistentAutoAttack class. Enhanced logging in InitialInformationPacketProcessor. Updated repository methods for async/await and null checks.

+++ src/Source/Distribution/DigitalWorldOnline.Game.Host/DungeonsServer/DungeonsServerBaseOperation.cs
@@ -398,6 +398,7 @@ namespace DigitalWorldOnline.GameHost
                 _logger.Error($"Unexpected error at map running: {ex.Message} {ex.StackTrace}.");
             }
         }
+    
 
 
         /// <summary>
@@ -409,8 +410,8 @@ namespace DigitalWorldOnline.GameHost
 
             if (client?.Tamer == null)
             {
-                _ =GameLogger.LogInfo($"Client sin Tamer detectado (AccountId: {client?.AccountId}). Forzando RemoveClient.");
-                    
+                _ = GameLogger.LogInfo($"Client sin Tamer detectado (AccountId: {client?.AccountId}). Forzando RemoveClient.");
+
                 RemoveClient(client);
                 return;
             }
@@ -825,7 +826,7 @@ namespace DigitalWorldOnline.GameHost
 
             return targetMobs;
         }
-        public IMob? GetIMobByHandler(short mapId,int handler,long tamerId)
+        public IMob? GetIMobByHandler(long mapId,int handler,long tamerId)
         {
             var map = Maps.FirstOrDefault(x => x.Clients.Exists(gameClient => gameClient.TamerId == tamerId));
